% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddShadedRegion.R
\name{AddShadedRegion}
\alias{AddShadedRegion}
\title{Add Background Shaded Region to an dygraph}
\usage{
AddShadedRegion(dyGraph, highlight, color = "red")
}
\arguments{
\item{dyGraph}{A dygraph object to add shading to}

\item{color}{Color of shading. This can be of the form "#AABBCC" or "rgb(255,
100, 200)" or "yellow". Defaults to be red.}

\item{hightlight}{A logical vector - must be of the same length as the time of
\code{dyGraph}}
}
\value{
A dygraph with the specified shading
}
\description{
Specify the region of a \code{dyGraph}, this function will add a
 background shading of \code{color} at specified region
}
\examples{
set.seed(1)
p <- PlotTimeSeries(ProcessorTime, "time", c("BN2", "CO4"))
highlight <- sample(c(T, F), length(ProcessorTime$time), replace = TRUE, prob = c(0.1, 0.9))
AddShadedRegion(p, highlight)
}
