% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetStartEndIndices.R
\name{GetStartEndIndices}
\alias{GetStartEndIndices}
\title{Indices of Continuous Subsequences of Certain Value}
\usage{
GetStartEndIndices(x, value = TRUE)
}
\arguments{
\item{x}{a vector of any given length and mode, e.g. "logical", "integer",
"numeric" etc, from which to find the index of \code{value}}

\item{value}{an element to search for}
}
\value{
A list of two vectors \itemize{\item \code{start} contains the start indices of continuous subsequeces of
 \code{value} inside vector \code{x} \item \code{end} contains the end indices of continuous subsequeces of
 \code{value} inside vector \code{x}}
}
\description{
Give all start and end indices of continuous subsequeces of
 \code{value} inside vector \code{x}
}
\examples{
x <- c(T, T)
GetStartEndIndices(x, T)
x <- c(F, F, T, F, T, T, T, F)
GetStartEndIndices(x, T)
x <- c(1, 1, 2, 3, 4, 1, 1, 1)
GetStartEndIndices(x, 1)
}
