% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LastPointAboveQuantile.R
\name{LastPointAboveQuantile}
\alias{LastPointAboveQuantile}
\title{Determine if the Last Point is Above the Sample Quantile}
\usage{
LastPointAboveQuantile(ts, exclude = NULL, probT = 0.9)
}
\arguments{
\item{ts}{A numeric vector whose \itemize{ \item last value is the one will
be compared with the sample quantile; \item all but the last one are used
to calculate the sample quantile.}}

\item{exclude}{A logical vector with length equals to \code{length(ts)-1}. It
is used to remove elements at designated positions from calculating the
sample quantile. By default, \code{exclude = NULL}, which means no element
is excluded when calculating the sample quantile.}

\item{probT}{probability threshold with values in \emph{[0, 1]}. The sample
quantile corresponds to this probability is calculated.}
}
\value{
returns \code{TRUE} if the last element of \code{ts} is above the
  sample quantiles and \code{FALSE} otherwise.
}
\description{
This function \code{LastPointAboveQuantile} is used to determine
  whether the last observation of \code{ts} is above the sample quantile
  calculated using the rest of the points in \code{ts}. If argument
  \code{exclude} is specified, elements at those designated positions are
  further removed from the sample quantile calculation.
}
\examples{
set.seed(1)
ts <- c(runif(1000), 0.95)
exclude <- sample(c(T, F), 1000, replace = T, prob = c(0.005, 0.995))
LastPointAboveQuantile(ts, exclude)
LastPointAboveQuantile(ts, exclude, 0.95)
}
\seealso{
\code{\link[stats]{quantile}}
}
