% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LastPointVSQuantile.R
\name{LastPointVSQuantile}
\alias{LastPointVSQuantile}
\title{Compare the Last Point with the Sample Quantile}
\usage{
LastPointVSQuantile(ts, exclude = NULL, probT = 0.9, direction = "pos")
}
\arguments{
\item{ts}{A numeric vector whose \itemize{ \item last value is the one will
be compared with the sample quantile; \item all but the last one are used
to calculate the sample quantile.}}

\item{exclude}{A logical vector with length equals to \code{length(ts)-1}. It
is used to remove elements at designated positions from calculating the
sample quantile. By default, \code{exclude = NULL}, which means no element
is excluded when calculating the sample quantile.}

\item{probT}{probability threshold with values in \emph{[0, 1]}. The sample
quantile corresponds to this probability is calculated. Depends on the
\code{direction}, if \itemize{\item \code{direction = "pos"} or
\code{direction = "neg"} \code{probT} should be a numeric value; \item
\code{direction = "both"} \code{prob} should be a numeric vector of two
values in \emph{[0, 1]}}.}

\item{direction}{Directionality of the anomalies to be deteted. Options
are: \emph{'pos'}, \emph{'neg'} and \emph{'both'}. Defaults to be
\emph{'pos'}.}
}
\value{
returns \code{TRUE} if the last element of \code{ts} is above or below the
  sample quantiles, \code{FALSE} otherwise.
}
\description{
This function \code{LastPointVSQuantile} is used to determine
  the relationship between the last observation of \code{ts} and the sample
  quantile calculated using the rest of the points in \code{ts}. If argument
  \code{exclude} is specified, elements at those designated positions are
  further removed from the sample quantile calculation.
}
\examples{
set.seed(1)
ts <- c(rnorm(1000), 5)
exclude <- sample(c(T, F), 1000, replace = T, prob = c(0.005, 0.995))
LastPointVSQuantile(ts, exclude, 0.95, "pos")
LastPointVSQuantile(ts, exclude, 0.95, "neg")
LastPointVSQuantile(ts, exclude, c(0.1, 0.9), "both")
}
\seealso{
\code{\link[stats]{quantile}}
}
