% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeLevels.R
\name{MergeLevels}
\alias{MergeLevels}
\title{Signal Quantization}
\usage{
MergeLevels(signal, minDiff)
}
\arguments{
\item{signal}{a numeric vector.}

\item{minDiff}{a numeric value indicating the maximum round-off error}
}
\value{
The returned value is a list of two vectors: \itemize{\item signal a
  numeric vector specifying the output signal (with smaller set of levels);
  \item levels.index an integer vector specifying the index of all the
  observations that belongs to any level}
}
\description{
This function rounds the input \code{signal} with a large set of
  input levels (values that more than 1 observation have) to a countable
  smaller set. The quantization error - the difference between an input value
  and its quantized value - is smaller than \code{minDiff}.
}
\examples{
signal <- c(rep(-1.1, 10), rep(5, 10), rep(-1, 10), 3, rep(4, 10), rep(3.9, 10))
output <- MergeLevels(signal, 0.2)
ts <- data.table(input = signal, output = output$signal, time = (as.Date("2017-01-01") + seq_along(signal)))
PlotTimeSeries(ts, "time")
}
