% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessorTime_data.R
\docType{data}
\name{ProcessorTime}
\alias{ProcessorTime}
\title{Average Processor Time in 3 Clusters}
\format{A data.table with 1 time column and 3 value columns named after
  clusters.}
\usage{
ProcessorTime
}
\description{
The data was extracted from Perf Counter Data in Cosmos for three
  clusters, i.e. \emph{BN2}, \emph{CH1B}, \emph{CO4}. It spans from \emph{2015-10-21} to
  \emph{2016-10-21} and gives the measurements in percentage of the variable
  \emph{processor time}.
}
\examples{
ProcessorTime
}
\references{
\href{http://sharepoint/sites/autopilot/default.aspx}{Perf Counter Data in Cosmos}
}
\keyword{datasets}
