% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SessionSuccessRate_data.R
\docType{data}
\name{SessionSuccessRate}
\alias{SessionSuccessRate}
\title{Session Success Rate}
\format{A data.table with 1 Day column in format YYYY-mm-dd and 1 value
  column.}
\usage{
SessionSuccessRate
}
\description{
Session success rate data for Bing that spans from
  \emph{2015-01-10} to \emph{2015-09-10}. This data contains no missing
  value.
}
\examples{
SessionSuccessRate
}
\keyword{datasets}
