% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPointIsAnomaly.R
\name{TestPointIsAnomaly}
\alias{TestPointIsAnomaly}
\title{Check Whether Test Point is Anomaly}
\usage{
TestPointIsAnomaly(training, test, dist = "t", exclude = NULL, p = 0.01,
  direction = direction, ...)
}
\arguments{
\item{training}{A numeric vector containing the samples used to fit the
distribution}

\item{test}{A numeric value to be tested}

\item{dist}{A string specifies the distribution to fit. Options are
t-distribution(default: \code{dist = "t"}), normal distribution(\code{dist
= "normal"}).}

\item{exclude}{A logical vector with length equals to
\code{length(training)}. It is used to remove elements at designated
positions from fitting the distribution. By default, \code{exclude = NULL},
which means no element is excluded.}

\item{p}{p-value threshold with values in \emph{[0, 1]}.}

\item{direction}{Directionality of the anomalies to be deteted. Options are:
\emph{'pos'}, \emph{'neg'} and \emph{'both'}. Defaults to be \emph{'both'}.}
}
\value{
returns \code{TRUE} if the test point is likely to be an anomaly and
  \code{FALSE} otherwise. For debugging purpose, this function also returns
  metadata \code{median}, \code{mad}, \code{score} and
  \code{hist}.\itemize{\item \code{median} sample median of training set
  \item \code{mad} sample mean absolute deviation calculated from training
  set \item \code{score} defined as \code{(test - median) / mad} \item
  \code{hist} histogram of training points overlaid with stats function used
  to fit}
}
\description{
Assume the \code{training} set was generated by a process that
  follows certain distribution, i.e. "normal" or "t", this function checks
  whether to reject the null hypothesis that the \code{test} point are
  generated by the same process. If the probability of obtaining a result
  equals to or more extreme than \code{test} is lower than \code{p}, then
  this function returns \code{TRUE}, meaning the null hypothesis is rejected
  and the \code{test} point is likely to be an anomoly. If argument
  \code{exclude} is specified, elements at those designated positions are
  removed from the training set.
}
\examples{
set.seed(1)
training <- rt(1000, df = 10)
test <- 4
exclude <- sample(c(T, F), 1000, replace = T, prob = c(0.005, 0.995))
r <- TestPointIsAnomaly(training, test, dist = "t", exclude = exclude, p = 0.1, df = 10)
attr(r, "hist")
training <- rnorm(1000, mean = 0, sd = 10)
r <- TestPointIsAnomaly(training, test, dist = "normal", p = 0.1)
attr(r, "hist")
}
